Fix make_directory calls on Unix systems
On Unix systems, creating a directory with a permission mask/mode of 0 prevents
any writes/reads to those directories. When that happens, the program is unable to write
the documentation files.

This commit changes the code to set the correct mode for Unix (Darwin, Linux, FreeBSD) and
keep using zero for all others.