Fix headers overlapping content above
This commit overrides the margin-top and padding-top css rules that were set for
heading elements in main content. The removed values probably were added to
prevent the heading element from going under the header element. However, this
was causing the heading elements to expand beyond their top over to the
preceding content, rendering the links and text inaccessible.