Escape less than in doc strings
There is a problem with core:encoding/xml, where the comments include
a lot of < characters, which are not escaped to &lt; and the result is
that the static site is not properly rendered for this package. I am
not sure but I hope that this change will escape < in the comments, so
that this problem is fixed.